Reorganize Python SDK clients to match Rust SDK structure
Move obligation clients to obligations/ subfolder and arbiter clients to
arbiters/ subfolder. Add nested module structure with property-based
access using #[getter] attributes for a more Pythonic API:
client.erc20.escrow.non_tierable.create() instead of method chaining.
