+++ sdks/py/src/clients/mod.rs
@@ -1,7 +1,7 @@
-pub mod attestation;
-pub mod erc1155;
-pub mod erc20;
-pub mod erc721;
-pub mod oracle;
-pub mod string_obligation;
-pub mod token_bundle;
+//! Clients module
+//!
+//! - `obligations/`: Token and attestation obligation clients
+//! - `arbiters/`: Arbiter clients for escrow arbitration
+
+pub mod arbiters;
+pub mod obligations;
