+++ sdks/py/src/clients/obligations/attestation.rs
