+++ sdks/py/src/clients/obligations/erc1155/barter_utils.rs
@@ -1,158 +1,43 @@
-use alkahest_rs::extensions::TokenBundleModule;
+//! ERC1155 barter utilities
+
+use alkahest_rs::extensions::Erc1155Module;
 use pyo3::{pyclass, pymethods, PyResult};
 
 use crate::{
     error_handling::{map_eyre_to_pyerr, map_parse_to_pyerr},
     get_attested_event,
-    types::{ArbiterData, AttestedLog, LogWithHash, TokenBundleData},
+    types::{AttestedLog, Erc1155Data, Erc20Data, Erc721Data, LogWithHash, TokenBundleData},
 };
 
+/// Barter utilities API for ERC1155 tokens
 #[pyclass]
 #[derive(Clone)]
-pub struct TokenBundleClient {
-    inner: TokenBundleModule,
+pub struct BarterUtils {
+    inner: Erc1155Module,
 }
 
-impl TokenBundleClient {
-    pub fn new(inner: TokenBundleModule) -> Self {
+impl BarterUtils {
+    pub fn new(inner: Erc1155Module) -> Self {
         Self { inner }
     }
 }
 
 #[pymethods]
-impl TokenBundleClient {
-    pub fn approve<'py>(
+impl BarterUtils {
+    // --- ERC1155 for ERC1155 ---
+
+    pub fn buy_erc1155_for_erc1155<'py>(
         &self,
         py: pyo3::Python<'py>,
-        token: TokenBundleData,
-        purpose: String,
-    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
-        let inner = self.inner.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let purpose = match purpose.as_str() {
-                "payment" => alkahest_rs::types::ApprovalPurpose::Payment,
-                "escrow" => alkahest_rs::types::ApprovalPurpose::Escrow,
-                _ => return Err(map_eyre_to_pyerr(eyre::eyre!("Invalid purpose"))),
-            };
-            let receipts = inner
-                .util()
-                .approve(&token.try_into().map_err(map_eyre_to_pyerr)?, purpose)
-                .await
-                .map_err(map_eyre_to_pyerr)?;
-
-            // Return the transaction hash of the last receipt, or empty string if no receipts
-            match receipts.last() {
-                Some(receipt) => Ok(receipt.transaction_hash.to_string()),
-                None => Ok("".to_string()),
-            }
-        })
-    }
-
-    pub fn collect_escrow<'py>(
-        &self,
-        py: pyo3::Python<'py>,
-        buy_attestation: String,
-        fulfillment: String,
-    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
-        let inner = self.inner.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let receipt = inner
-                .escrow()
-                .non_tierable()
-                .collect(
-                    buy_attestation.parse().map_err(map_parse_to_pyerr)?,
-                    fulfillment.parse().map_err(map_parse_to_pyerr)?,
-                )
-                .await
-                .map_err(map_eyre_to_pyerr)?;
-            Ok(receipt.transaction_hash.to_string())
-        })
-    }
-
-    pub fn reclaim_expired<'py>(
-        &self,
-        py: pyo3::Python<'py>,
-        buy_attestation: String,
-    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
-        let inner = self.inner.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let receipt = inner
-                .escrow()
-                .non_tierable()
-                .reclaim_expired(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
-                .await
-                .map_err(map_eyre_to_pyerr)?;
-            Ok(receipt.transaction_hash.to_string())
-        })
-    }
-
-    pub fn buy_with_bundle<'py>(
-        &self,
-        py: pyo3::Python<'py>,
-        price: TokenBundleData,
-        item: ArbiterData,
-        expiration: u64,
-    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
-        let inner = self.inner.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let receipt = inner
-                .escrow()
-                .non_tierable()
-                .create(
-                    &price.try_into().map_err(map_eyre_to_pyerr)?,
-                    &item.try_into().map_err(map_eyre_to_pyerr)?,
-                    expiration,
-                )
-                .await
-                .map_err(map_eyre_to_pyerr)?;
-            Ok(LogWithHash::<AttestedLog> {
-                log: get_attested_event(receipt.clone())
-                    .map_err(map_eyre_to_pyerr)?
-                    .data
-                    .into(),
-                transaction_hash: receipt.transaction_hash.to_string(),
-            })
-        })
-    }
-
-    pub fn pay_with_bundle<'py>(
-        &self,
-        py: pyo3::Python<'py>,
-        price: TokenBundleData,
-        payee: String,
-    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
-        let inner = self.inner.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let receipt = inner
-                .payment()
-                .pay(
-                    &price.try_into().map_err(map_eyre_to_pyerr)?,
-                    payee.parse().map_err(map_parse_to_pyerr)?,
-                )
-                .await
-                .map_err(map_eyre_to_pyerr)?;
-            Ok(LogWithHash::<AttestedLog> {
-                log: get_attested_event(receipt.clone())
-                    .map_err(map_eyre_to_pyerr)?
-                    .data
-                    .into(),
-                transaction_hash: receipt.transaction_hash.to_string(),
-            })
-        })
-    }
-
-    pub fn buy_bundle_for_bundle<'py>(
-        &self,
-        py: pyo3::Python<'py>,
-        bid: TokenBundleData,
-        ask: TokenBundleData,
+        bid: Erc1155Data,
+        ask: Erc1155Data,
         expiration: u64,
     ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
         let inner = self.inner.clone();
         pyo3_async_runtimes::tokio::future_into_py(py, async move {
             let receipt = inner
                 .barter()
-                .buy_bundle_for_bundle(
+                .buy_erc1155_for_erc1155(
                     &bid.try_into().map_err(map_eyre_to_pyerr)?,
                     &ask.try_into().map_err(map_eyre_to_pyerr)?,
                     expiration,
@@ -169,7 +54,7 @@ impl TokenBundleClient {
         })
     }
 
-    pub fn pay_bundle_for_bundle<'py>(
+    pub fn pay_erc1155_for_erc1155<'py>(
         &self,
         py: pyo3::Python<'py>,
         buy_attestation: String,
@@ -178,7 +63,163 @@ impl TokenBundleClient {
         pyo3_async_runtimes::tokio::future_into_py(py, async move {
             let receipt = inner
                 .barter()
-                .pay_bundle_for_bundle(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
+                .pay_erc1155_for_erc1155(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    // --- ERC1155 for ERC20 ---
+
+    pub fn buy_erc20_with_erc1155<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        bid: Erc1155Data,
+        ask: Erc20Data,
+        expiration: u64,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .buy_erc20_with_erc1155(
+                    &bid.try_into().map_err(map_eyre_to_pyerr)?,
+                    &ask.try_into().map_err(map_eyre_to_pyerr)?,
+                    expiration,
+                )
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    pub fn pay_erc1155_for_erc20<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        buy_attestation: String,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .pay_erc1155_for_erc20(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    // --- ERC1155 for ERC721 ---
+
+    pub fn buy_erc721_with_erc1155<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        bid: Erc1155Data,
+        ask: Erc721Data,
+        expiration: u64,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .buy_erc721_with_erc1155(
+                    &bid.try_into().map_err(map_eyre_to_pyerr)?,
+                    &ask.try_into().map_err(map_eyre_to_pyerr)?,
+                    expiration,
+                )
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    pub fn pay_erc1155_for_erc721<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        buy_attestation: String,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .pay_erc1155_for_erc721(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    // --- ERC1155 for TokenBundle ---
+
+    pub fn buy_bundle_with_erc1155<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        bid: Erc1155Data,
+        ask: TokenBundleData,
+        expiration: u64,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .buy_bundle_with_erc1155(
+                    &bid.try_into().map_err(map_eyre_to_pyerr)?,
+                    &ask.try_into().map_err(map_eyre_to_pyerr)?,
+                    expiration,
+                )
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    pub fn pay_erc1155_for_bundle<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        buy_attestation: String,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .pay_erc1155_for_bundle(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
                 .await
                 .map_err(map_eyre_to_pyerr)?;
             Ok(LogWithHash::<AttestedLog> {
