+++ sdks/py/src/clients/obligations/string_obligation.rs
