+++ sdks/py/src/clients/obligations/token_bundle/barter_utils.rs
@@ -1,0 +1,78 @@
+//! TokenBundle barter utilities
+
+use alkahest_rs::extensions::TokenBundleModule;
+use pyo3::{pyclass, pymethods, PyResult};
+
+use crate::{
+    error_handling::{map_eyre_to_pyerr, map_parse_to_pyerr},
+    get_attested_event,
+    types::{AttestedLog, LogWithHash, TokenBundleData},
+};
+
+/// Barter utilities API for TokenBundle
+#[pyclass]
+#[derive(Clone)]
+pub struct BarterUtils {
+    inner: TokenBundleModule,
+}
+
+impl BarterUtils {
+    pub fn new(inner: TokenBundleModule) -> Self {
+        Self { inner }
+    }
+}
+
+#[pymethods]
+impl BarterUtils {
+    /// Creates an escrow to buy one token bundle for another.
+    pub fn buy_bundle_for_bundle<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        bid: TokenBundleData,
+        ask: TokenBundleData,
+        expiration: u64,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .buy_bundle_for_bundle(
+                    &bid.try_into().map_err(map_eyre_to_pyerr)?,
+                    &ask.try_into().map_err(map_eyre_to_pyerr)?,
+                    expiration,
+                )
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+
+    /// Fulfills a bundle-for-bundle trade.
+    pub fn pay_bundle_for_bundle<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        buy_attestation: String,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .barter()
+                .pay_bundle_for_bundle(buy_attestation.parse().map_err(map_parse_to_pyerr)?)
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+}
