+++ sdks/py/src/clients/obligations/token_bundle/payment.rs
@@ -1,0 +1,53 @@
+//! TokenBundle payment obligation client
+
+use alkahest_rs::extensions::TokenBundleModule;
+use pyo3::{pyclass, pymethods, PyResult};
+
+use crate::{
+    error_handling::{map_eyre_to_pyerr, map_parse_to_pyerr},
+    get_attested_event,
+    types::{AttestedLog, LogWithHash, TokenBundleData},
+};
+
+/// Payment API for TokenBundle
+#[pyclass]
+#[derive(Clone)]
+pub struct Payment {
+    inner: TokenBundleModule,
+}
+
+impl Payment {
+    pub fn new(inner: TokenBundleModule) -> Self {
+        Self { inner }
+    }
+}
+
+#[pymethods]
+impl Payment {
+    /// Makes a direct payment with a token bundle.
+    pub fn pay<'py>(
+        &self,
+        py: pyo3::Python<'py>,
+        price: TokenBundleData,
+        payee: String,
+    ) -> PyResult<pyo3::Bound<'py, pyo3::PyAny>> {
+        let inner = self.inner.clone();
+        pyo3_async_runtimes::tokio::future_into_py(py, async move {
+            let receipt = inner
+                .payment()
+                .pay(
+                    &price.try_into().map_err(map_eyre_to_pyerr)?,
+                    payee.parse().map_err(map_parse_to_pyerr)?,
+                )
+                .await
+                .map_err(map_eyre_to_pyerr)?;
+            Ok(LogWithHash::<AttestedLog> {
+                log: get_attested_event(receipt.clone())
+                    .map_err(map_eyre_to_pyerr)?
+                    .data
+                    .into(),
+                transaction_hash: receipt.transaction_hash.to_string(),
+            })
+        })
+    }
+}
